+++ server/src/main/java/org/opensearch/search/aggregations/bucket/terms/GlobalOrdinalsStringTermsAggregator.java
@@ -207,28 +207,21 @@ public class GlobalOrdinalsStringTermsAggregator extends AbstractStringTermsAggr
             return false;
         }
 
-        TermsEnum indexTermsEnum = segmentTerms.iterator();
-        BytesRef indexTerm = indexTermsEnum.next();
-        final SortedSetDocValues globalOrds = this.getGlobalOrds(ctx);
-        TermsEnum globalOrdinalTermsEnum = globalOrds.termsEnum();
-        BytesRef ordinalTerm = globalOrdinalTermsEnum.next();
+        final TermsEnum segmentTermsEnum = segmentTerms.iterator();
+        final LongUnaryOperator globalOrdsMapping = valuesSource.globalOrdinalsMapping(ctx);
 
-        // Iterate over the terms in the segment, look for matches in the global ordinal terms,
-        // and increment bucket count when segment terms match global ordinal terms.
-        while (indexTerm != null && ordinalTerm != null) {
-            int compare = indexTerm.compareTo(ordinalTerm);
-            if (compare == 0) {
-                if (acceptedGlobalOrdinals.test(globalOrdinalTermsEnum.ord())) {
-                    ordCountConsumer.accept(globalOrdinalTermsEnum.ord(), indexTermsEnum.docFreq());
-                }
-                indexTerm = indexTermsEnum.next();
-                ordinalTerm = globalOrdinalTermsEnum.next();
-            } else if (compare < 0) {
-                indexTerm = indexTermsEnum.next();
-            } else {
-                ordinalTerm = globalOrdinalTermsEnum.next();
+        // Iterate over the ordinals in the segment, look for matches in the global ordinal,
+        // and increment bucket count when segment ordinal is contained in global ordinals.
+        for (long segmentOrd = 0; segmentOrd < termCount; segmentOrd++) {
+            segmentTermsEnum.next();
+            long globalOrd = globalOrdsMapping.applyAsLong(segmentOrd);
+            if (acceptedGlobalOrdinals.test(globalOrd)) {
+                ordCountConsumer.accept(globalOrd, segmentTermsEnum.docFreq());
             }
         }
+
+        // Segment terms enum should be fully exhausted after iterating over all the ordinals
+        assert segmentTermsEnum.next() == null : "Segment terms enum must be fully exhausted";
         return true;
     }
 
@@ -565,11 +558,10 @@ public class GlobalOrdinalsStringTermsAggregator extends AbstractStringTermsAggr
         @Override
         protected boolean tryPrecomputeAggregationForLeaf(LeafReaderContext ctx) throws IOException {
             if (subAggregators.length == 0) {
-                if (mapping != null) {
-                    mapSegmentCountsToGlobalCounts(mapping);
-                }
-                mapping = valuesSource.globalOrdinalsMapping(ctx);
-                return tryCollectFromTermFrequencies(ctx, (ord, docCount) -> incrementBucketDocCount(mapping.applyAsLong(ord), docCount));
+                return tryCollectFromTermFrequencies(
+                    ctx,
+                    (globalOrd, docCount) -> incrementBucketDocCount(collectionStrategy.globalOrdToBucketOrd(0, globalOrd), docCount)
+                );
             }
             return tryStarTreePrecompute(ctx);
         }